+++ packages/terra-application/src/application-status-overlay/ApplicationStatusOverlayProvider.jsx
@@ -21,9 +21,16 @@ const propTypes = {
    * ApplicationStatusOverlayProvider.
    */
   scrollRefCallback: PropTypes.func,
+  /**
+   * @private
+   * Prop to not render scroll in  Powerchart.
+   */
+  noScroll: PropTypes.bool,
 };
 
-const ApplicationStatusOverlayProvider = ({ children, scrollRefCallback, ...customProps }) => {
+const ApplicationStatusOverlayProvider = ({
+  children, scrollRefCallback, noScroll, ...customProps
+}) => {
   const [registeredStatusOverlay, setRegisteredStatusOverlay] = React.useState({});
 
   const containerRef = React.useRef();
@@ -101,11 +108,13 @@ const ApplicationStatusOverlayProvider = ({ children, scrollRefCallback, ...cust
     />
   );
 
+  const scrollClass = cx({ 'remove-scroll': noScroll });
+
   return (
     <div {...customProps} className={className}>
       {statusView}
       <div data-status-overlay-container-content ref={containerRef} className={cx('container-content')}>
-        <Scroll refCallback={scrollRefCallback}>
+        <Scroll className={scrollClass} refCallback={scrollRefCallback}>
           <ApplicationStatusOverlayContext.Provider value={contextValue}>
             {children}
           </ApplicationStatusOverlayContext.Provider>